requirements/requirements.txt
# coding:utf-8
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

Django==2.1.7
#django-compressor==2.2
django-cors-headers==2.1.0
PyMySQL==0.9.2
requests==2.18.4
pymongo==3.6.1
ua-parser==0.8.0
PyExecJS==1.5.1
#pandas==0.23.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-bulk==0.2.1
gunicorn==19.3.0
psutil==5.4.6
PyYAML==3.13
#zmail==0.2.0
filelock==2.0.13
paramiko==2.4.2
sshpubkeys==3.1.0
#itsdangerous
#ipdb
#werkzeug==0.15.2
xmltodict
celery==4.1.0
kombu==4.0.2
flower==0.9.3
redis==2.10.6
django-redis==4.9.0
django-redis-sessions==0.6.1
django-celery-beat==1.4.0
eventlet==0.24.1
python-libnmap==0.7.0
pysnmp==4.4.9

feedparser==5.2.1
git+git://github.com/sshwsfc/xadmin.git@django2